+++ app/src/main/java/com/flint/core/common/di/RepositoryModule.kt
@@ -2,8 +2,8 @@ package com.flint.core.common.di
 
 import com.flint.data.repositoryImpl.DefaultAuthRepository
 import com.flint.data.repositoryImpl.DefaultUserRepository
-import com.flint.domain.repository.AuthRepository
-import com.flint.domain.repository.UserRepository
+import com.flint.data.repository.AuthRepository
+import com.flint.data.repository.UserRepository
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
