+++ app/src/main/java/com/flint/core/designsystem/component/listItem/SavedContentItem.kt
@@ -18,7 +18,7 @@ import com.flint.core.common.extension.noRippleClickable
 import com.flint.core.designsystem.component.image.NetworkImage
 import com.flint.core.designsystem.component.listView.OttHorizontalList
 import com.flint.core.designsystem.theme.FlintTheme
-import com.flint.domain.model.ContentModel
+import com.flint.data.model.content.ContentModel
 import com.flint.domain.type.OttType
 
 @Composable
