+++ app/src/main/java/com/flint/data/api/FlintApi.kt
@@ -1,8 +1,8 @@
 package com.flint.data.api
 
 import com.flint.data.dto.base.BaseResponse
-import com.flint.data.dto.request.SampleRequestDto
-import com.flint.data.dto.response.SampleResponseDto
+import com.flint.data.dto.SampleRequestDto
+import com.flint.data.dto.SampleResponseDto
 import retrofit2.http.Body
 import retrofit2.http.GET
