+++ app/src/main/java/com/flint/data/dto/SampleResponseDto.kt
@@ -1,4 +1,4 @@
-package com.flint.data.dto.response
+package com.flint.data.dto
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
