+++ app/src/main/java/com/flint/data/dto/auth/request/SocialVerifyRequestDto.kt
@@ -1,12 +1,12 @@
-package com.flint.data.dto.request.auth
+package com.flint.data.dto.auth.request
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class SocialVerifyRequestDto (
+data class SocialVerifyRequestDto(
     @SerialName("provider")
-    val provider : String,
+    val provider: String,
     @SerialName("accessToken")
-    val accessToken : String,
+    val accessToken: String,
 )
