+++ app/src/main/java/com/flint/data/dto/auth/response/SignupResponseDto.kt
@@ -1,14 +1,14 @@
-package com.flint.data.dto.response.auth
+package com.flint.data.dto.auth.response
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class SignupResponseDto (
+data class SignupResponseDto(
     @SerialName("accessToken")
-    val accessToken : String,
+    val accessToken: String,
     @SerialName("refreshToken")
-    val refreshToken : String,
+    val refreshToken: String,
     @SerialName("userId")
-    val userId : Long,
+    val userId: Long,
 )
