+++ app/src/main/java/com/flint/data/mapper/SampleMapper.kt
@@ -1,6 +1,6 @@
 package com.flint.data.mapper
 
-import com.flint.data.dto.response.SampleResponseDto
+import com.flint.data.dto.SampleResponseDto
 import com.flint.domain.model.SampleModel
 
 fun SampleResponseDto.toModel(): SampleModel =
