+++ app/src/main/java/com/flint/data/repository/AuthRepository.kt
@@ -1,4 +1,4 @@
-package com.flint.domain.repository
+package com.flint.data.repository
 
 import com.flint.data.model.auth.SignupRequestModel
 import com.flint.data.model.auth.SignupResponseModel
@@ -7,5 +7,6 @@ import com.flint.data.model.auth.SocialVerifyResponseModel
 
 interface AuthRepository {
     suspend fun signup(requestModel: SignupRequestModel): Result<SignupResponseModel>
+
     suspend fun socialVerify(requestModel: SocialVerifyRequestModel): Result<SocialVerifyResponseModel>
 }
