+++ app/src/main/java/com/flint/data/repositoryImpl/DefaultFlintRepository.kt
@@ -1,10 +1,10 @@
 package com.flint.data.repositoryImpl
 
 import com.flint.data.api.FlintApi
-import com.flint.data.dto.request.SampleRequestDto
+import com.flint.data.dto.SampleRequestDto
 import com.flint.data.mapper.toModel
 import com.flint.domain.model.SampleModel
-import com.flint.domain.repository.FlintRepository
+import com.flint.data.repository.FlintRepository
 import javax.inject.Inject
 
 class DefaultFlintRepository
