--- app/src/main/java/com/flint/data/util/SafeApiCall.kt
@@ -1,8 +1,0 @@
-package com.flint.data.util
-
-inline fun <T> safeApiCall(block: () -> T): Result<T> =
-    runCatching { block() }
-        .fold(
-            onSuccess = { Result.success(it) },
-            onFailure = { Result.failure(it) },
-        )
