--- app/src/main/java/com/flint/domain/model/OttModel.kt
@@ -1,7 +1,0 @@
-package com.flint.domain.model
-
-data class OttModel(
-    val ottId: Long,
-    val ottName: String,
-    val logoUrl: String,
-)
